Move LayerManager to managers package
